+++ ninja-core/src/main/java/ninja/params/Params.java
@@ -29,7 +29,12 @@ import java.lang.annotation.Target;
  * @author James Moger
  *
  */
-@WithArgumentExtractor(ArgumentExtractors.ParamsExtractor.class)
+@WithArgumentExtractors({
+    ArgumentExtractors.ParamsExtractor.class,
+    ArgumentExtractors.FileItemParamsExtractor.class,
+    ArgumentExtractors.FileParamsExtractor.class,
+    ArgumentExtractors.InputStreamParamsExtractor.class
+})
 @Retention(RetentionPolicy.RUNTIME)
 @Target({ElementType.PARAMETER})
 public @interface Params {