+++ ninja-core/src/site/markdown/developer/changelog.md
@@ -1,3 +1,9 @@
+Version x.x.x
+=============
+
+ * 2015-06-24 Injection of params and uploaded files from multipart requests (bazi, momiji)
+
+
 Version 5.1.4
 =============